test(ai): Rename auth gateway integration helper
Use a purpose-based helper name for the shared auth-gateway integration harness used by users, pools, management, and audit tests.

Refs #5098
